implement adding comments

+++ app/controllers/articles_controller.rb
@@ -9,6 +9,8 @@ class ArticlesController < ApplicationController
 
   def show
     @article = Article.find(params[:id])
+    @comment = @article.comments.build
+    @comments = @article.comments
   end
 
   def new
@@ -51,9 +53,14 @@ class ArticlesController < ApplicationController
  end
 
  def destroy
-  if @article.destroy
-    flash[:success] = "Article has been deleted."
-    redirect_to articles_path
+  unless @article.user == current_user
+    flash[:danger] = "You can only delete your own article."
+    redirect_to root_path 
+  else
+    if @article.destroy
+      flash[:success] = "Article has been deleted."
+      redirect_to articles_path
+    end
   end
 end
